+++ packages/gamut-labs/src/Notifications/NotificationsPopover.tsx
@@ -1,22 +1,29 @@
 import { FlexBox, Popover } from '@codecademy/gamut';
 import React from 'react';
 
+import { CrossDeviceItemId } from '../GlobalHeader/types';
 import { NotificationsContents } from './NotificationsContents';
 import { NotificationsRendererProps } from './types';
 
 export const NotificationsPopover: React.FC<NotificationsRendererProps> = ({
   bellRef,
-  onClose,
+  setOpenCrossDeviceItemId,
   ...props
 }) => {
+  const onRequestCloseHandler = () => {
+    // When on the mobile view, we don't want the desktop handler to fire.
+    if (window.innerWidth >= 1200) {
+      setOpenCrossDeviceItemId?.(CrossDeviceItemId.UNSET);
+    }
+  };
   return (
     <Popover
       align="right"
       verticalOffset={1}
       outline
       isOpen
-      onRequestClose={onClose}
       targetRef={bellRef}
+      onRequestClose={onRequestCloseHandler}
     >
       <FlexBox bg="white" textAlign="left" width={360}>
         <NotificationsContents {...props} />