+++ packages/gamut-labs/src/Notifications/__tests__/useHeaderNotifications-test.tsx
@@ -5,12 +5,13 @@ import userEvent from '@testing-library/user-event';
 import { times } from 'lodash';
 import React from 'react';
 
+import { CrossDeviceItemId } from '../../GlobalHeader/types';
 import { createStubNotification } from '../__fixtures__/stubs';
 import { useHeaderNotifications } from '../useHeaderNotifications';
 
 const onEnable = jest.fn();
 
-const defaultProps = {
+const settingsProps = {
   actions: {
     clear: jest.fn(),
     click: jest.fn(),
@@ -21,21 +22,26 @@ const defaultProps = {
   onEnable,
 };
 
+const defaultProps = {
+  settings: settingsProps,
+  Renderer: () => <div>hi!</div>,
+  openCrossDeviceItemId: CrossDeviceItemId.UNSET,
+  setOpenCrossDeviceItemId: jest.fn(),
+};
+
 describe('useHeaderNotifications', () => {
   it('returns nulls when there are no notifications', async () => {
     const hook = renderHook(() =>
-      useHeaderNotifications(undefined, () => <div>hi!</div>)
+      useHeaderNotifications({ ...defaultProps, settings: undefined })
     );
-
     expect(hook.result.current).toEqual([null, null]);
   });
 
   it('renders its notifications pane as invisible by default when there are notifications', async () => {
     const notifications = [createStubNotification()];
+    const newSettings = { ...settingsProps, notifications };
     const hook = renderHook(() =>
-      useHeaderNotifications({ ...defaultProps, notifications }, () => (
-        <div>hi!</div>
-      ))
+      useHeaderNotifications({ ...defaultProps, settings: newSettings })
     );
 
     const view = render(
@@ -46,13 +52,16 @@ describe('useHeaderNotifications', () => {
   });
 
   it('renders its notifications pane as visible when the bell is clicked and there are notifications', async () => {
+    const mockSetOpenCrossDeviceItemId = jest.fn();
+    defaultProps.setOpenCrossDeviceItemId = mockSetOpenCrossDeviceItemId;
+
     const notifications = times(5, (id) =>
       createStubNotification({ id: `${id}`, unread: true })
     );
+    const newSettings = { ...settingsProps, notifications };
+
     const hook = renderHook(() =>
-      useHeaderNotifications({ ...defaultProps, notifications }, () => (
-        <div>hi!</div>
-      ))
+      useHeaderNotifications({ ...defaultProps, settings: newSettings })
     );
 
     const buttonView = render(
@@ -65,11 +74,28 @@ describe('useHeaderNotifications', () => {
       userEvent.click(buttonView.getByRole('menuitem'));
     });
 
+    // verify button click would tell react to update state
+    expect(mockSetOpenCrossDeviceItemId).toBeCalledTimes(1);
+    expect(mockSetOpenCrossDeviceItemId).toBeCalledWith('notifications');
+
+    const overriddenDefaultProps = {
+      ...defaultProps,
+      openCrossDeviceItemId: CrossDeviceItemId.NOTIFICATIONS,
+    };
+    // simulate React telling the component there was a change to its prop (openCrossDeviceItemId) and therefore trigger a rerender
+    const secondRenderOfHook = renderHook(() =>
+      useHeaderNotifications({
+        ...overriddenDefaultProps,
+        settings: newSettings,
+      })
+    );
     const paneView = render(
-      <MockGamutProvider>{hook.result.current[1]}</MockGamutProvider>
+      <MockGamutProvider>
+        {secondRenderOfHook.result.current[1]}
+      </MockGamutProvider>
     );
 
-    expect(defaultProps.onEnable).toHaveBeenCalled();
+    expect(settingsProps.onEnable).toHaveBeenCalled();
     expect(paneView.container).not.toBeEmptyDOMElement();
   });
 });