+++ packages/gamut-labs/src/Notifications/types.ts
@@ -8,7 +8,7 @@ export type NotificationActions = {
   track: (target: string) => void;
 };
 
-export type AppHeaderNotifications = {
+export type AppHeaderNotificationSettings = {
   actions: NotificationActions;
   notifications: Notification[];
   onEnable: () => void;
@@ -18,9 +18,9 @@ export type NotificationsContentsProps = {
   actions: NotificationActions;
   notifications: Notification[];
   notificationListRef?: React.RefObject<HTMLDivElement>;
+  setOpenCrossDeviceItemId?: React.Dispatch<React.SetStateAction<string>>;
 };
 
 export type NotificationsRendererProps = NotificationsContentsProps & {
   bellRef: React.RefObject<ButtonBaseElements>;
-  onClose: () => void;
 };