+++ packages/gamut/src/List/types.ts
@@ -1,7 +1,8 @@
 export interface PrivateListProps {
   scrollable?: boolean;
-  spacing?: 'normal' | 'condensed';
+  spacing?: 'normal' | 'condensed' | 'compact';
   variant?: 'default' | 'table' | 'card' | 'block' | 'plain';
+  rowBreakpoint?: 'xs' | 'sm' | 'md';
 }
 
 export type PublicListProps<T> = Omit<T, keyof PrivateListProps>;